+++ CHANGELOG.rst
@@ -35,6 +35,8 @@ Added
   Contributed by @amanda11
 * Ensure `.pth` files in the st2 virtualenv get loaded by pack virtualenvs. #6183
   Contributed by @cognifloyd
+* Add support for SSL/TLS to `st2api` and `st2stream` components. #6204
+  Contributed by @jk464
 
 3.8.1 - December 13, 2023
 -------------------------